# openapi.yaml
openapi: 3.0.3
info:
  title: Marketplace API
  description: |
    A RESTful marketplace API with Users, Items, and Orders.
    
    ## Authentication
    Use the **Authorize** button (🔒) at the top to paste your JWT token as:
    `Bearer <your_token_here>`
    
    ## Seed Credentials
    | User  | Email                  | Password     | Role  |
    |-------|------------------------|--------------|-------|
    | Admin | admin@example.com      | Admin123!    | ADMIN |
    | Alice | alice@example.com      | Password123! | USER  |
    | Bob   | bob@example.com        | Password123! | USER  |
  version: 1.0.0

servers:
  - url: /api
    description: API base path

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id: { type: integer, example: 1 }
        email: { type: string, example: alice@example.com }
        name: { type: string, example: Alice Smith }
        role: { type: string, enum: [USER, ADMIN] }
        createdAt: { type: string, format: date-time }

    Item:
      type: object
      properties:
        id: { type: integer, example: 1 }
        title: { type: string, example: Laptop }
        description: { type: string, example: A powerful laptop }
        price: { type: number, example: 1200.00 }
        quantity: { type: integer, example: 5 }
        sellerId: { type: integer, example: 2 }
        seller:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
            email: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Order:
      type: object
      properties:
        id: { type: integer, example: 1 }
        status:
          type: string
          enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
        total: { type: number, example: 1450.00 }
        buyerId: { type: integer, example: 3 }
        buyer:
          type: object
          properties:
            id: { type: integer }
            name: { type: string }
            email: { type: string }
        orderItems:
          type: array
          items:
            type: object
            properties:
              id: { type: integer }
              quantity: { type: integer }
              price: { type: number }
              item:
                type: object
                properties:
                  id: { type: integer }
                  title: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    Error:
      type: object
      properties:
        error: { type: string }

tags:
  - name: Auth
    description: Signup and login
  - name: Users
    description: User management
  - name: Items
    description: Marketplace items
  - name: Orders
    description: Orders and order management

paths:
  # ── AUTH ──────────────────────────────────────────────────────────────────
  /auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, name]
              properties:
                email: { type: string, example: newuser@example.com }
                password: { type: string, example: Password123! }
                name: { type: string, example: New User }
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '400':
          description: Missing required fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Login and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, example: alice@example.com }
                password: { type: string, example: Password123! }
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  token: { type: string }
        '400':
          description: Missing fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── USERS ─────────────────────────────────────────────────────────────────
  /users:
    get:
      tags: [Users]
      summary: List all users (Admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /users/{id}:
    get:
      tags: [Users]
      summary: Get a user by ID (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 2 }
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Invalid ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      tags: [Users]
      summary: Update a user (Admin or self)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 2 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, example: Updated Name }
                email: { type: string, example: updated@example.com }
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: No fields provided
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409':
          description: Email already in use
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [Users]
      summary: Delete a user (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 2 }
      responses:
        '200':
          description: User deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── ITEMS ─────────────────────────────────────────────────────────────────
  /items:
    get:
      tags: [Items]
      summary: List all items (Public)
      responses:
        '200':
          description: List of items
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Item' }

    post:
      tags: [Items]
      summary: Create an item (Authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, price]
              properties:
                title: { type: string, example: Gaming Mouse }
                description: { type: string, example: High-DPI gaming mouse }
                price: { type: number, example: 79.99 }
                quantity: { type: integer, example: 10 }
      responses:
        '201':
          description: Item created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400':
          description: Missing required fields
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /items/{id}:
    get:
      tags: [Items]
      summary: Get an item by ID (Public)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Item found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400':
          description: Invalid ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      tags: [Items]
      summary: Update an item (Owner or Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string, example: Updated Laptop }
                description: { type: string }
                price: { type: number, example: 1100.00 }
                quantity: { type: integer, example: 3 }
      responses:
        '200':
          description: Item updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Item' }
        '400':
          description: No fields / invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden – not the seller
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [Items]
      summary: Delete an item (Owner or Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Item deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  # ── ORDERS ────────────────────────────────────────────────────────────────
  /orders:
    get:
      tags: [Orders]
      summary: List orders (Admin sees all; Users see own)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Order' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    post:
      tags: [Orders]
      summary: Place an order (Authenticated)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [items]
              properties:
                items:
                  type: array
                  items:
                    type: object
                    required: [itemId, quantity]
                    properties:
                      itemId: { type: integer, example: 1 }
                      quantity: { type: integer, example: 1 }
            example:
              items:
                - itemId: 1
                  quantity: 1
                - itemId: 2
                  quantity: 2
      responses:
        '201':
          description: Order placed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Bad request (missing items, insufficient stock)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Item not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order by ID (Owner or Admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Invalid ID
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    put:
      tags: [Orders]
      summary: Update order status (Admin any status; Owner cancel only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [PENDING, CONFIRMED, SHIPPED, DELIVERED, CANCELLED]
                  example: SHIPPED
      responses:
        '200':
          description: Order updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400':
          description: Invalid status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }

    delete:
      tags: [Orders]
      summary: Delete an order (Admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, example: 1 }
      responses:
        '200':
          description: Order deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404':
          description: Order not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }